Adds warning about using discharge with clamshell
Without reading the issues, I enabled battery discharge on the commandline (while using the laptop in clamshell mode). I had to jump through a few hoops to get clamshell mode working again. Thus, probably best to warn users not to use the `battery discharge` command if they are using it in clamshell mode.